Got hardware interrrupts working.
The issue before leading to only one irq firing and no irq firing
after software interrupt was that 'sti' was not called at the end of
the interrupt handlers.

The code now needs cleaning up for consistency. The next major task is
to write a keyboard driver. Note the timer handler does nothing - an
internal clock has not been implemented.