Keyboard driver works for keys implemented
Not all keys implemented - TODO includes shift and backspace.

Got it to work by implementing the array as a single string with
multiple break points on the stack, thus avoiding the use of both
arrays and the data segment which were previously causing problems.

Would be better to switch string to .rodata in the future though so
it doesn't need to be re-defined each time (not that there is any
noticable lag).