Added comments to various files for GutHub upload
Nothing changed other than commenting. Code tested to work as before.

The timer drivers are still inactive. A next step is to complete them,
and either modify the screen driver so that the time elapsed can be
constantly shown at the top of the screen, or wait untill there is
a shell to call the function.