Timer driver working with new interface
timer.c_ignore and timer.h_ignore files are now called timer.c and
timer.h. They contain the code for the timer driver, which is called
by os/cpu/interrupts.c. The old 'dots printed on screen' timer in
interrupts.c was removed. The new timer driver prints to time
elapsed in form HH MM SS to the top of the screen, which remains upon
scrolling. This required modification of os/driver/screen.c (and .h)
to start scrolling on the second line, and for the addition of the
printHeader() function. The kernel was modified to reflect the change
of the welcome message form printAt() to printHeader().

Makefile and link.ld were modified to compile with the new timer driver
c code files.

The bootloader at os/boot/bootloader.asm was also modified to load 20
rather than 9 pages to memory. This fixed a bug in which every key
printed 'S '.