Minor changes in preparing for paging
The bootloader int 0x13 will load a maximum of 54 pages into memory
(for unknown reason). This means that the paging tables cannot be
stored on disk even for 1GiB mapping. The plan is to dynamically
create the tables from physical address 0x100000. Would be preferable
to have it the other way round with the kernel at the higher address
after the paging tables, but can only access addresses up to 0xffff in
read mode.

Removed errenous comment in os/boot/read_sectors_16bit.asm.