Basic paging setup but needs much work
Paging setup (not PAE or PSE) and working with first 4MiB physical memory
mapped to 0.75GiB for kernel and identity mapped to 0GiB, in files
paging.asm and kernel_entry.asm. However, when the identity mapping is
unmapped (as is desirable for user space programs to run there), the OS
triple faults and reboots. I believe this is because the GDTR no longer
addresses the GDT (as set up with physical address) and so IDT doesn't
work (remember requires segment address). A GDT is to be implemented in
C in os/cpu (needed for switch to userspace anyway).

Other files that were modified to enable paging were:
 - Screen driver at os/drivers/screen.c had direct memory mapped text
   mode video buffer address changed from physical address 0xb8000 to
   virtual address 0x300b8000 (0xb8000 + 0.75GiB).
 - Makefile had addition of new file paging.asm.
 - Linker script needed to be updated with new starting address of
   kernel, and new file.

Next steps:
 - Implement new GDT in physical address space
 - Unmap identity mapping
 - Clean up paging.asm ('global' declarations not needed as '%include'd?)
 - Check comments on new files (inc. timer driver)