Added unique handlers for software interrupts 0-31
This change revealed that the OS page faults (interrupt 14) if the OS
is left to run too long (around 2 mins). This particularly seems to
occur if a key is pressed passed this point. The cause of this bug is
unknown but may be to do with the GDT.

To aid in future OS development, a list of pitfalls and bugs come
accross may be created.