Fixed bug where OS would page fault after ~2mins
The bug was because the stack was initialized using the .bss
segment, which itself is not initialized. This was fixed by changing
the stack address to an absolute (virtual) address (0x30300000 chosen).

The OS was tested to run for over 20 mins without any errors.

Further research is to be done to understand the cause of the bug, see
next commit(s). Also, in trying to fix the bug uneccessary code/
comments added which will be removed.