Paging code cleaned up after bug fix.
Redundant code and comments in paging section removed. Regarding the
bug, setting the stack to size 0x4000 using uninitialized .bss set the
stack start address to (virtual) address 0x4000. This address is listed
in the page directory entry (PDE) as being not present, thus explaining
the page fault (though the exact timing reason is unknown, but is
dependent on stack size).