Cleaned up drivers codes, fixed cursor bug
No major changes to logic, functions simply shortened and comments
altered for readability.

Bug found where cursor not visible on final line after scroll. This
was because the line was cleared by filling it with zeros, which
prevented the cursor from being shown. Replacing the zeros with spaces
(and attribute bytes of DEF_STYLE) solved this issue.