Improved Cross compiler support and Makefile
Directory for cross compiler has been mooved within the project
directory where it is more convienient. The instructions in the
README on how to install it have been improoved. The files
themselves (except the README) have been added to the gitignore.

The Makefile has been updated and improoved such that it refuses
to work if the cross compiler of mkisofs dependencies are not
present.