Makefile now size-independent and re-runnable.
Previously, 'make fail' or 'make clean' (as applicable) had to be
run before make could be run again, as the object files etc. created
were moved after there creation to a separate directory (.bin/). Now
all files created by Makefile are created in this directory from the
start. This, and general other cleaning up (e.g. of 'make clean'), means
that 'make' can be called multiple times without fail. The 'make fail'
function is now redundant and has been removed.

More significantly, the Makefile has been modified such that it does
not need to increase with the size of the project. Instead of files or
directories being listed individually as they are made, this process
is now automatic. As linker scripts do not support variables, the
'INPUT' section was mooved to the calling recipe in the Makefile.

TODO: Modify README to account for these changes.