Moved kernel load to 0x9000 so no bootloader clash
Previously the kernel was loaded at 0x1000. However, this only
allowed 54 pages to be loaded by the bootloader as the 55 page
would be loaded at address 0x7c00, overwriting the bootloader
itself. When this was done, the OS simply stopped doing anything,
producing no error message. Moving the loading address to 0x9000
avoids this confusion.

The paging tables were moved to load from 0x19000, allowing 128
pages to be loaded by the BIOS. Furthermore, it is now easy to
change this, as only the paging need be changed.

These changes were complemented with a new 'memory addresses'
section in the README.

Trivial change to os/boot/gdt.asm removed unneccessary link also
present in README.

TODO: add accout of problem in pitfalls.txt.