Makefile checks for repeat file names
Repeat files names prevent compilation due to a limitation
of the Makefile. This is not currently a major constraint
and simpilifies the compilation process. Consequently a section
has simply been added to the Makefile that checks for repeat
file names (base names, e.g. os/kernel/a.c and os/drivers/a.asm
would count as repeats of 'a') and stops compilation if any
found.