Added GPLv3 boilerplate to all code files
Boilerplate text deviates slightly from that suggested by
GNU for the sake of brevity. However it retains all the meaning
and clarity of the original suggestion, which could equally have
been used.

For more details of the license, a copy of the GNU General Public
Licence Version 3 (GPLv3) as published by the Free Software Foundation
should have been distributed with this software (under filename 'LICENSE').
If not, see <http://www.gnu.org/licenses/>.