Fixed BIOS 'int 0x13' address overflow issue
When the address to load the BIOS to was moved from 0x1000
to 0x9000 to avoid overriting the bootloader at 0x7c00
(see prev commit), it was overlooked that the BIOS READ
INT 0x13 function is in 16-bit real mode. As such, whilst
it worked on Qemu, it failed to boot on real hardware as
it was asked to access RAM addresses above 0xffff. This
failed silently, with no error code reported.