Added printf-like 'printk' to screen driver
'printk' function converts '%x' to unsigned hexadecimal in the
same way as printf does (and is based off the printf code).
'%%' converts to print '%' also as in printf. A '%' followed
by anything else prints literally without modification.

This was created to better understand the ATAPI driver, which
works in Qemu but does not give the expected return values on
bare metal. These values can now be printed.

TODO:
 - clean up screen driver code
 - finish ATAPI driver
 - make it work with iso9660 file system