Added baskspace to keyboard and screen drivers
Backspace was decided to be an essential feature. However, other
keys, such as shift, are still not implemented. The reason for this
is that this project is largely intended to be fun and educational,
and such unneccessary polishes would add to the code base without
contributing to any greater understanding of OSs.