IDE ATAPI dectection driver written
Driver correctly identifies ATAPI device when run on Qemu, executing
exactly as expected. However, it produces an error upon reading the
status when run on actual hardware (command/status register: 0x7f),
and other registers are 0xff. No devices are found in spite of the
PC having an IDE controller and a SATA controller according to
'lspci' in linux. This may be because the ATAPI device is modelled
as being directly attached in Qemu, but is on the PCI bus on the
actual PC.