Cleaned up ATAPI and added PCI drivers
Further research indicates the ATAPI driver should have recognised
the optical drive on my bare metal testing PC. The drive was
successfully located by PCI driver. Given the impossibility of
dealing with specifics of each different hardware manufactorer,
and the code bloat it would cause with little added value
it was deciced to instead test on VirtualBox. The ATAPI and PCI
drivers work as expected on VirtualBox.

The PCI driver is thus not currently being used, but has been kept
for future use. If nothing else it can be used for an 'lspci'
implementation.

ioports files modified for addition of 32-bit IO port usage
required for PCI.